Improve objseq.ObjSeq_UpdateCurvePosition 99.36% -> 100.00%
ObjSeqState::groundSnapEnabled is a script-toggled 0/1 flag that retail
passes straight to RomCurveInterp_EvaluateOffsetPosition's int param
(lbz, no extsb); the s8 spelling forced a spurious sign-extension. Type it
u8 - the field is only referenced from objseq.c, and every read already
spells an explicit (s8) cast, which is a no-op unless the field is u8.

The one uncast read, in ObjSeq_update, needs that same (s8) cast to keep
retail's extsb/cmpwi form; it now matches the (s8) spelling the sibling
sites (ObjSeq_SetupInitialPlaybackState, ObjSeq_ToggleCommand3Target)
already use. Both edits are required together: the u8 retype alone regresses
the unit to 99.71770, and the cast alone is inert.

+++ src/main/objseq.c
@@ -4970,7 +4970,7 @@ int ObjSeq_update(u8* obj, f32 t)
             ((ObjAnimBank*)action)->currentState->eventCountdown = (u16)(int)(lbl_803DF050 * state->fade);
         }
         ObjSeq_UpdateCurvePosition(obj, seq);
-        if (state->groundSnapEnabled == 1 &&
+        if ((s8)state->groundSnapEnabled == 1 &&
             hitDetectFn_800658a4((GameObject*)obj, ((GameObject*)obj)->anim.localPosX, ((GameObject*)obj)->anim.localPosY,
                                  ((GameObject*)obj)->anim.localPosZ, scratch, 0) == 0)
         {
